notification_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.30
botocore==1.37.30
click==8.1.8
fastapi==0.115.12
h11==0.14.0
idna==3.10
jmespath==1.0.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
